+++ contracts/vesting-lp-pcl/Cargo.toml
@@ -18,9 +18,10 @@ library = []
 cw2 = { workspace = true }
 cw20 = { workspace = true }
 astroport = { workspace = true }
-vesting-base-pcl = { workspace = true }
+vesting-base = {path = "../../packages/vesting-base"}
 cosmwasm-schema = { workspace = true }
 cosmwasm-std = { workspace = true }
+cw-storage-plus = "0.15"
 
 [dev-dependencies]
 cw-multi-test = { workspace = true }
