+++ contracts/vesting-lp-pcl/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod contract;
 pub mod msg;
+pub mod state;
 
 #[cfg(test)]
 mod tests;
