+++ contracts/vesting-lp/Cargo.toml
@@ -15,11 +15,12 @@ library = []
 [dependencies]
 cw2 = { version = "0.15" }
 vesting-base = {path = "../../packages/vesting-base"}
-vesting-base-lp = {path = "../../packages/vesting-base-lp"}
+vesting-lp-pcl = {path = "../vesting-lp-pcl"}
 astroport = { git = "https://github.com/astroport-fi/astroport-core.git", tag = "v2.8.0" }
 cosmwasm-schema = { version = "1.1", default-features = false  }
 cosmwasm-std = { version = "1.1" }
 cw-storage-plus = "0.15"
+cw20 = { version = "0.15" }
 
 [dev-dependencies]
 cw-multi-test = "0.15"
