+++ contracts/vesting-lp/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod contract;
 pub mod msg;
+pub mod state;
