--- packages/vesting-base-lp/Cargo.toml
@@ -1,26 +1,0 @@
-[package]
-name = "vesting-base-lp"
-version = "1.1.0"
-authors = ["Astroport"]
-edition = "2021"
-
-[lib]
-crate-type = ["cdylib", "rlib"]
-
-[features]
-backtraces = ["cosmwasm-std/backtraces"]
-# use library feature to disable all init/handle/query exports
-library = []
-
-[dependencies]
-serde = { version = "1.0.145", default-features = false, features = ["derive"] }
-cw2 = { version = "0.15" }
-cw20 = { version = "0.15" }
-cosmwasm-std = { version = "1.1" }
-cw-storage-plus = "0.15"
-thiserror = { version = "1.0" }
-cw-utils = "0.15"
-cosmwasm-schema = { version = "1.1", default-features = false  }
-vesting-base = {path = "../vesting-base"}
-vesting-base-pcl = {path = "../vesting-base-pcl"}
-astroport = { git = "https://github.com/astroport-fi/astroport-core.git", tag = "v2.8.0" }
