--- packages/vesting-base-lp/src/lib.rs
@@ -1,10 +1,0 @@
-pub mod builder;
-pub mod error;
-pub mod handlers;
-pub mod msg;
-pub mod state;
-pub mod types;
-
-pub(crate) mod ext_historical;
-pub(crate) mod ext_managed;
-pub(crate) mod ext_with_managers;
