--- packages/vesting-base-pcl/Cargo.toml
@@ -1,23 +1,0 @@
-[package]
-name = "vesting-base-pcl"
-version = "1.1.0"
-authors = ["Astroport"]
-edition = "2021"
-
-[lib]
-crate-type = ["cdylib", "rlib"]
-
-[features]
-backtraces = ["cosmwasm-std/backtraces"]
-# use library feature to disable all init/handle/query exports
-library = []
-
-[dependencies]
-cw20 = { workspace = true }
-cosmwasm-std = { workspace = true }
-cw-storage-plus = { workspace = true }
-astroport = { workspace = true }
-thiserror = { workspace = true }
-# we keep it at 0.15 instead of latest version just for vesting investors contract
-cw-utils = "0.15"
-cosmwasm-schema = { workspace = true }
